+++ var/www/api/get2.php
@@ -38,6 +38,7 @@ header("Access-Control-Allow-Origin: *");
 // define("ROOT", "/home/ubuntu/Development/zboota-server"); // Development ROOT
 require_once dirname(__FILE__).'/../../../config.php';
 require_once ROOT.'/lib/getCore.php';
+require_once ROOT.'/lib/errorInLpns.php';
 
 if(isset($argc) && $argc>=2) {
 	$lpns=$argv[1];
@@ -61,6 +62,18 @@ if($lpns=="") {
 	return;
 }
 
+// decode json
 $lpns=json_decode($lpns,true);
+
+// verify
+$eilpns=errorInLpns($lpns);
+if(!!$eilpns) {
+	echo json_encode(array('error'=>$eilpns);
+	return;
+}
+
+// get data
 $data=getCore($lpns,$force);
+
+// return json data
 echo json_encode($data);